switch to new core
