+++ plugins/platform-core-old/babel.config.js
@@ -3,5 +3,8 @@ module.exports = {
   presets: [
     ['@babel/preset-env', { targets: { node: 'current' } }],
     '@babel/preset-typescript',
+  ],
+  plugins: [
+    ["@babel/plugin-proposal-decorators", { "legacy": true }],
   ]
 }
