+++ plugins/platform-core-old/jest.config.js
@@ -7,5 +7,4 @@ module.exports = {
   roots: [
     '<rootDir>/src'
   ],
-  testMatch: ["**/?(*.)+(spec|test).[jt]s?(x)"]
 }
