+++ plugins/platform-core-old/src/__resources__/builder.ts
