+++ plugins/platform-core-old/src/__resources__/index.ts
