+++ plugins/platform-core-old/src/__resources__/model.ts
