+++ plugins/platform-core-old/src/__resources__/strings/ru.ts
