+++ plugins/platform-core-old/src/__resources__/utils.ts
