+++ plugins/platform-core-old/tsconfig.json
