+++ plugins/platform-core/babel.config.js
@@ -3,8 +3,5 @@ module.exports = {
   presets: [
     ['@babel/preset-env', { targets: { node: 'current' } }],
     '@babel/preset-typescript',
-  ],
-  plugins: [
-    ["@babel/plugin-proposal-decorators", { "legacy": true }],
   ]
 }
