+++ plugins/platform-core/jest.config.js
@@ -7,4 +7,5 @@ module.exports = {
   roots: [
     '<rootDir>/src'
   ],
+  testMatch: ["**/?(*.)+(spec|test).[jt]s?(x)"]
 }
