+++ plugins/platform-core/src/__tests__/core.test.ts
