+++ plugins/platform-core/src/plugin.ts
@@ -1,238 +1,213 @@
 //
 // Copyright © 2020 Anticrm Platform Contributors.
-//
+// 
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
 // obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//
+// 
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
 
-import { Platform, Resource, AnyPlugin } from '@anticrm/platform'
-import { Db } from '@anticrm/platform-db'
+import { Platform } from '@anticrm/platform'
 import core, {
-  Obj, Doc, Ref, Bag, Class, Type, Emb,
-  PropertyType, BagOf, Content, CorePlugin
+  CoreService, Obj, Ref, Class, Doc, EClass, BagOf, InstanceOf,
+  Instance, Type, Emb, ResourceType, Property, ResourceProperty
 } from '.'
-import { TSession, SessionProto, Konstructor, Layout } from './session'
 
-/// ///////
+export default async (platform: Platform) => {
 
-console.log('PLUGIN: parsed core')
+  type Konstructor<T extends Obj> = new (obj: Omit<T, '__property' | '_class'>) => Instance<T>
 
-/*!
- * Anticrm Platform™ Core Plugin
- * © 2020 Anticrm Platform Contributors. All Rights Reserved.
- * Licensed under the Eclipse Public License, Version 2.0
- */
-export default async (platform: Platform, deps: { db: Db }): Promise<CorePlugin> => {
-  console.log('PLUGIN: started core')
-
-  abstract class TSessionProto implements SessionProto {
-    getSession (): TSession { throw new Error('session provide the implementation') }
-    abstract __mapKey (_class: Ref<Class<Obj>>, key: string): string | null
+  interface InstanceProxy {
+    __layout: any
   }
 
-  class TEmb extends TSessionProto implements Emb {
-    _class!: Ref<Class<this>>
-    toIntlString (plural?: number): string { return this.getClass().toIntlString(plural) }
-    getClass (): Class<this> {
-      return this.getSession().getInstanceSync(this._class)
-    }
-
-    __mapKey (_class: Ref<Class<Obj>>, key: string) { return key }
-  }
-
-  class TDoc extends TSessionProto implements Doc {
-    _class!: Ref<Class<this>>
-    _id!: Ref<this>
-    toIntlString (plural?: number): string { return this.getClass().toIntlString(plural) }
-    getClass (): Class<this> {
-      return this.getSession().getInstanceSync(this._class)
-    }
-
-    as<T extends Doc> (_class: Ref<Class<T>>): Promise<T | undefined> {
-      return this.getSession().as(this as unknown as Layout<Doc>, _class)
-    }
-
-    mixins (): Ref<Class<Doc>>[] {
-      const layout = this as unknown as Layout<Doc>
-      return layout.__layout._classes as Ref<Class<Doc>>[]
-    }
-
-    __mapKey (_class: Ref<Class<Obj>>, key: string) { return key.startsWith('_') ? key : _class + '/' + key }
-  }
-
-  // T Y P E S
-
-  class TType<T extends PropertyType> extends TEmb implements Type<T> {
-    _default?: T
-    exert (value: T, target?: PropertyType, key?: PropertyKey): any { return value ?? this._default } // eslint-disable-line
-    hibernate (value: any): T { return value }
-  }
-
-  class TInstanceOf<T extends Emb> extends TType<T> {
-    of!: Ref<Class<T>>
-    exert (value: T) {
-      if (typeof value === 'object') { return this.getSession().instantiateSync(value._class, value) }
-      return undefined
+  const funcs = {
+    identity: function (value: Property<any>): any {
+      console.log('!!!!!! IDENTITY')
+      return value
+    },
+    instanceOf: function (value: Property<any>): any {
+      console.log('!!!!!! INSTANCEOF')
+      return value
+    },
+    bagOf: function (value: Property<any>): any {
+      console.log('!!!!!! BAGOF')
+      return value
     }
   }
 
-  // C O L L E C T I O N S : A R R A Y
+  // const ResourceType_exert = 
 
-  class ArrayProxyHandler implements ProxyHandler<PropertyType[]> {
-    private type: Type<PropertyType>
+  const objects = new Map<Ref<Doc>, Doc>()
+  const byClass = new Map<Ref<Class<Doc>>, Doc[]>()
 
-    constructor (type: Type<PropertyType>) {
-      this.type = type
-    }
 
-    get (target: PropertyType[], key: PropertyKey): any {
-      const value = Reflect.get(target, key)
-      return this.type.exert(value)
-    }
+  function get<T extends Doc> (_id: Ref<T>): T {
+    const result = objects.get(_id)
+    if (result) { return result as T }
+    throw new Error('oops! object not found: ' + _id)
   }
 
-  class TArrayOf<T extends PropertyType> extends TType<T[]> {
-    of!: Type<T>
-    exert (value: T[]) {
-      console.log('array')
-      console.log(value)
-      return new Proxy(value, new ArrayProxyHandler(this.of))
-    }
+  // C L A S S E S
+
+  function getOwnAttribute (clazz: Class<Obj>, key: string): Type<any> | undefined {
+    return (clazz._attributes as any)[key] ??
+      (clazz._overrides ? (clazz._overrides as any)[key] : undefined)
   }
 
-  // C O L L E C T I O N S : B A G
-
-  class BagProxyHandler implements ProxyHandler<Bag<PropertyType>> {
-    private type: Type<PropertyType>
-
-    constructor (type: Type<PropertyType>) {
-      this.type = type
-    }
-
-    get (target: Bag<PropertyType>, key: string): any {
-      const value = Reflect.get(target, key)
-      return this.type.exert(value)
-    }
+  function getAttribute (clazz: Class<Obj>, key: string): Type<any> | undefined {
+    return getOwnAttribute(clazz, key) ??
+      (clazz._extends ? getAttribute(get(clazz._extends), key) : undefined)
   }
 
-  class TBagOf<T extends PropertyType> extends TType<Bag<T>> implements BagOf<T> {
-    of!: Type<T>
-    exert (value: Bag<T>) {
-      if (typeof value === 'object') { return new Proxy(value, new BagProxyHandler(this.of)) }
-      return undefined
-    }
+  // I N S T A N T I A T I O N
+
+  const konstructors = new Map<Ref<Class<Obj>>, Konstructor<Obj>>()
+  const prototypes = new Map<Ref<Class<Obj>>, Object>()
+
+  const CoreRoot = {
+    get _class (this: Instance<Obj>) { return this.__layout._class }
   }
 
-  // S T R U C T U R A L  F E A T U R E S
+  function getPrototype<T extends Obj> (_class: Ref<Class<T>>): Object {
+    const prototype = prototypes.get(_class)
+    if (prototype) {
+      return prototype
+    }
 
-  abstract class TStructuralFeature<T extends Obj> extends TDoc implements Class<T> {
-    _attributes!: Bag<Type<PropertyType>>
-    _extends?: Ref<Class<Obj>>
-    _native?: Resource<T>
+    const clazz = get(_class) as Class<Obj>
+    const parent = clazz._extends ? getPrototype(clazz._extends) : CoreRoot
+    const proto = Object.create(parent)
+    prototypes.set(_class, proto)
 
-    abstract createConstructor (): Konstructor<T>
+    if (_class as string === core.class.ResourceType) {
+      proto.exert = function (this: Instance<ResourceType<any>>, value: Property<any>): any {
+        const funcName = (value ?? this.__layout._default) as ResourceProperty<() => any>
 
-    newInstance (data: Content<T>): Promise<T> {
-      const session = this.getSession()
-      let ctor = session.constructors.get(this._id) as Konstructor<T>
-      if (!ctor) {
-        ctor = this.createConstructor()
-        session.constructors.set(this._id, ctor)
+        const f = platform.getResource(funcName)
+        if (f) return f
+
+        const func = (funcs as any)[funcName]
+        if (!func)
+          throw new Error('no resourcetype: ' + funcName)
+        return func
       }
-      return ctor(data)
-    }
-  }
-
-  class TStruct<T extends Obj> extends TStructuralFeature<T> {
-    toIntlString (plural?: number): string { return 'struct: ' + this._id } // eslint-disable-line
-
-    createConstructor (): Konstructor<T> {
-      const session = this.getSession()
-      const _class = this._id as Ref<Class<T>>
-      return data => session.createEmb(_class, data)
-    }
-  }
-
-  class TClass<T extends Doc> extends TStructuralFeature<T> {
-    toIntlString (plural?: number): string { return 'doc: ' + this._id } // eslint-disable-line
-
-    createConstructor (): Konstructor<T> {
-      const session = this.getSession()
-      const _class = this._id as Ref<Class<T>>
-      return data => session.createDoc(_class, data)
-    }
-  }
-
-  // B O O T  S E S S I O N  &  P L U G I N
-
-  class TCorePlugin implements CorePlugin {
-    readonly platform: Platform
-    readonly pluginId = core.id
-
-    private session: TSession
-    private prototypes = new Map<Resource<object>, object>()
-
-    constructor (platform: Platform, session: TSession) {
-      this.platform = platform
-      this.session = session
     }
 
-    async resolve (resource: Resource<object>): Promise<object> {
-      const proto = this.prototypes.get(resource)
-      if (proto) { return proto }
-      const index = resource.indexOf(':') + 1
-      const dot = resource.indexOf('.', index)
-      const plugin = resource.substring(index, dot) as AnyPlugin
-      return platform.getPlugin(plugin).then(plugin => {
-        const proto = this.prototypes.get(resource)
-        if (proto) { return proto }
-        throw new Error('plugin ' + plugin + ' does not provide resource: ' + resource)
+    const attributes: { [key: string]: Type<any> } = { ...clazz._attributes, ...clazz._overrides }
+    for (const key in attributes) {
+      if (key === '_default') { continue } // we do not define `_default`'s type, it's infinitevely recursive :)
+      const attr = attributes[key]
+      const attrInstance = instantiate(attr)
+
+      const exert = attrInstance.exert as (value: Property<any>) => any
+      if (typeof exert !== 'function') {
+        throw new Error('exert must be a function, ' + exert)
+      }
+
+      const bound = exert.bind(attrInstance)
+      Object.defineProperty(proto, key, {
+        get (this: InstanceProxy) {
+          return bound(this.__layout[key])
+        },
+        enumerable: true
       })
     }
+    return proto
+  }
 
-    registerPrototype<T extends Obj> (id: Resource<T>, proto: T): void {
-      if (this.prototypes.get(id)) { throw new Error('prototype ' + id + ' already registered') }
-      this.prototypes.set(id, proto)
-    }
-
-    getSession () { return this.session }
-
-    // U T I L I T Y
-
-    async getClassHierarchy (_class: Ref<Class<Obj>>): Promise<Ref<Class<Obj>>[]> {
-      const result = [] as Ref<Class<Obj>>[]
-      let clazz = _class as Ref<Class<Obj>> | undefined
-      while (clazz) {
-        result.push(clazz)
-        const instance = await this.getSession().getInstance(clazz)
-        clazz = instance._extends
-      }
-      return result.reverse()
+  function getKonstructor<T extends Obj> (_class: Ref<Class<T>>): Konstructor<T> {
+    const konstructor = konstructors.get(_class)
+    if (konstructor) { return konstructor as unknown as Konstructor<T> }
+    else {
+      // build ctor for _class
+      const proto = getPrototype(_class)
+      const ctor = {
+        [_class]: function (this: InstanceProxy, obj: Obj) {
+          this.__layout = obj
+        }
+      }[_class]
+      proto.constructor = ctor
+      ctor.prototype = proto
+      konstructors.set(_class, ctor as unknown as Konstructor<Obj>)
+      return ctor as unknown as Konstructor<T>
     }
   }
 
-  const session = new TSession(platform, deps.db)
-  const plugin = new TCorePlugin(platform, session)
+  function instantiate<T extends Obj> (obj: T): Instance<T> {
+    const ctor = getKonstructor(obj._class)
+    return new ctor(obj)
+  }
 
-  plugin.registerPrototype(core.native.Emb, TEmb.prototype)
-  plugin.registerPrototype(core.native.Doc, TDoc.prototype)
+  // C O R E  A P I
 
-  plugin.registerPrototype(core.native.Type, TType.prototype)
-  plugin.registerPrototype(core.native.BagOf, TBagOf.prototype)
-  plugin.registerPrototype(core.native.ArrayOf, TArrayOf.prototype)
-  plugin.registerPrototype(core.native.InstanceOf, TInstanceOf.prototype)
+  function mixin<D extends T, M extends T, T extends Doc> (doc: D, clazz: Ref<EClass<M, T>>, values: Pick<M, Exclude<keyof M, keyof T>>): M {
+    throw new Error("Method not implemented.")
+  }
 
-  plugin.registerPrototype(core.native.StructuralFeature, TStructuralFeature.prototype)
-  plugin.registerPrototype(core.native.Class, TClass.prototype)
-  plugin.registerPrototype(core.native.Struct, TStruct.prototype)
+  function newInstance<M extends Emb> (_class: Ref<Class<M>>, values: Omit<M, keyof Emb>): M {
+    const obj = { _class, ...values } as M
+    return obj
+  }
 
-  return plugin
+  function loadDocument<M extends Doc> (_class: Ref<Class<M>>, values: Omit<M, keyof Doc>): M {
+    const obj = { _class, ...values } as M
+    objects.set(obj._id, obj)
+
+    return obj
+  }
+
+  function newDocument<M extends Doc> (_class: Ref<Class<M>>, values: Omit<M, keyof Doc>): Instance<M> {
+    throw new Error("Method not implemented.")
+    //return this.getKonstructor(_class)(this.loadDocument(_class, values))
+  }
+
+  function loadClass<T extends E, E extends Obj> (values: Omit<EClass<T, E>, keyof Obj>): EClass<T, E> {
+    return loadDocument(core.class.Class as Ref<Class<EClass<T, E>>>, values)
+  }
+
+  function newClass<T extends E, E extends Obj> (values: Omit<EClass<T, E>, keyof Obj>): Instance<EClass<T, E>> {
+    return newDocument(core.class.Class as Ref<Class<EClass<T, E>>>, values)
+  }
+
+  // T Y P E S : B A G
+
+  class BagProxyHandler implements ProxyHandler<any> {
+    private exert: (value: Property<any>) => any
+
+    constructor(type: Instance<Type<any>>) {
+      if (!type.exert) {
+        throw new Error('bagof: no exert')
+      }
+      this.exert = type.exert
+    }
+
+    get (target: any, key: string): any {
+      return this.exert(Reflect.get(target, key))
+    }
+  }
+
+  const BagOf_exert = function (this: Instance<BagOf<any>>, value: { [key: string]: Property<any> }): { [key: string]: any } {
+    return new Proxy(value, new BagProxyHandler(this.of))
+  }
+
+  const InstanceOf_exert = function (this: Instance<InstanceOf<Emb>>, value: Emb): Instance<Emb> {
+    return instantiate(value)
+  }
+
+  platform.setResource(core.method.BagOf_exert, BagOf_exert)
+  platform.setResource(core.method.InstanceOf_exert, InstanceOf_exert)
+
+  return {
+    mixin, newInstance, loadDocument, newDocument, loadClass, newClass,
+
+    getPrototype, get, instantiate
+  }
 }
+
