--- plugins/platform-memdb/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@anticrm/platform-memdb",
-  "version": "0.1.0",
-  "main": "src/index.ts",
-  "author": "Andrey Platov <andrey.v.platov@gmail.com>",
-  "license": "EPL-2.0",
-  "scripts": {
-    "build": "tsc",
-    "test": "jest",
-    "doc": "typedoc --out ./generated_docs"
-  },
-  "devDependencies": {
-    "typedoc": "^0.17.6",
-    "typescript": "~3.8.3"
-  }
-}
